pyproject.toml
[project]
name = "invoice_validator"
version = "0.1.0"
description = "invoice-validator using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai-tools>=0.2.0",
    "crewai[google-genai,tools]>=0.28.0",
    "jsonlines>=4.0.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "litellm[proxy]>=1.75.3",
    "loguru>=0.7.0",
    "pandas>=2.0.0",
    "pillow>=10.0.0",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "pypdf2>=3.0.0",
    "pytesseract>=0.3.10",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "sqlalchemy>=2.0.0",
    "loguru>=0.7.0",
    "gradio>=4.0.0",
    "plotly>=5.0.0",
    "transformers>=4.57.6",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "pdf2image>=1.17.0",
    "pypdfium2>=4.24.0",
    "sentencepiece>=0.2.0",
]

[project.scripts]
invoice_validator = "invoice_validator.main:run"
run_crew = "invoice_validator.main:run"
train = "invoice_validator.main:train"
replay = "invoice_validator.main:replay"
test = "invoice_validator.main:test"
run_with_trigger = "invoice_validator.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
